save user bets, access condition

+++ src/java/ch/bfh/awebt/bookmaker/presentation/GameBean.java
@@ -2,6 +2,7 @@ package ch.bfh.awebt.bookmaker.presentation;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
@@ -13,10 +14,12 @@ import javax.annotation.PostConstruct;
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.ManagedProperty;
 import javax.faces.bean.ViewScoped;
+import javax.persistence.PersistenceException;
 import ch.bfh.awebt.bookmaker.persistence.BetDAO;
 import ch.bfh.awebt.bookmaker.persistence.GameDAO;
 import ch.bfh.awebt.bookmaker.persistence.TeamDAO;
 import ch.bfh.awebt.bookmaker.persistence.UserBetDAO;
+import ch.bfh.awebt.bookmaker.persistence.UserDAO;
 import ch.bfh.awebt.bookmaker.persistence.data.Bet;
 import ch.bfh.awebt.bookmaker.persistence.data.Game;
 import ch.bfh.awebt.bookmaker.persistence.data.Team;
@@ -40,15 +43,20 @@ public class GameBean implements Serializable {
 
 	private transient TeamDAO teamDAO;
 	private transient GameDAO gameDAO;
-	private transient UserBetDAO userBetDAO;
+	private transient UserDAO userDAO;
 	private transient BetDAO betDAO;
+	private transient UserBetDAO userBetDAO;
 
-	private Integer gameId;
-	private String gameTeam1;
-	private String gameTeam2;
+	private boolean statusConfirmation;
+
+	private Integer gameId, gameBetsUserId;
+	private String gameTeam1, gameTeam2;
 	private LocalDateTime gameStartTime;
 	private ZoneId gameTimeZone;
 	private List<BetDTO> gameBets;
+	private BigDecimal gameBetAmount, gamePaymentAmount;
+	private String gamePaymentCreditCardNumber, gamePaymentCreditCardCode;
+	private Integer gamePaymentCreditCardExpirationMonth, gamePaymentCreditCardExpirationYear, gamePaymentCreditCardExpirationYearMinimum, gamePaymentCreditCardExpirationYearMaximum;
 
 	/**
 	 * Initialises the managed bean.
@@ -95,6 +103,14 @@ public class GameBean implements Serializable {
 		return gameDAO;
 	}
 
+	private UserDAO getUserDAO() {
+
+		if (userDAO == null)
+			userDAO = new UserDAO();
+
+		return userDAO;
+	}
+
 	private BetDAO getBetDAO() {
 
 		if (betDAO == null)
@@ -109,6 +125,11 @@ public class GameBean implements Serializable {
 			userBetDAO = new UserBetDAO();
 
 		return userBetDAO;
+	}
+
+	public boolean isConfirmation() {
+
+		return statusConfirmation;
 	}
 
 	/**
@@ -127,8 +148,12 @@ public class GameBean implements Serializable {
 	 * @param gameId unique identifier of the game
 	 */
 	public void setGameId(Integer gameId) {
+		setGameId(gameId, false);
+	}
 
-		if (this.gameId == null || !this.gameId.equals(gameId)) {
+	private void setGameId(Integer gameId, boolean force) {
+
+		if (force || this.gameId == null || !this.gameId.equals(gameId)) {
 			this.gameId = gameId;
 
 			Game game = gameId != null ? getGame() : null;
@@ -138,12 +163,16 @@ public class GameBean implements Serializable {
 				gameStartTime = game.getStartTimeZoned().withZoneSameInstant(gameTimeZone).toLocalDateTime();
 
 				User u = loginBean.getUser();
+				gameBetsUserId = u != null ? u.getId() : null;
 				gameBets = new ArrayList<>(game.getBets()).stream().map(b -> { //BUGFIX: new ArrayList<>(...) needed in eclipselink < 2.7
 					UserBet c = u != null ? getUserBetDAO().findByUserAndBet(u, b) : null;
 					return new BetDTO(b, c != null ? c : new UserBet(u != null ? u : User.ANONYMOUS, b, BigDecimal.ZERO));
 				}).collect(Collectors.toList());
 
 			} else {
+				if (gameId != null)
+					MessageFactory.addWarning("ch.bfh.awebt.bookmaker.ITEM_NOT_FOUND_ERROR");
+
 				gameTeam1 = gameTeam2 = null;
 				gameStartTime = null;
 				gameBets = Arrays.asList();
@@ -229,7 +258,7 @@ public class GameBean implements Serializable {
 
 	public ZonedDateTime getGameStartTimeZoned() {
 
-		return ZonedDateTime.of(gameStartTime, gameTimeZone);
+		return gameStartTime != null ? ZonedDateTime.of(gameStartTime, gameTimeZone) : null;
 	}
 
 	/**
@@ -238,6 +267,10 @@ public class GameBean implements Serializable {
 	 * @return bets associated with the game
 	 */
 	public List<BetDTO> getGameBets() {
+
+		if (loginBean.getUserId() != null && !loginBean.getUserId().equals(gameBetsUserId))
+			setGameId(gameId, true);
+
 		return gameBets;
 	}
 
@@ -248,6 +281,132 @@ public class GameBean implements Serializable {
 	 */
 	public void setGameBets(List<BetDTO> gameBets) {
 		this.gameBets = gameBets;
+	}
+
+	/**
+	 * Gets the amount to bet.
+	 *
+	 * @return the amount to bet
+	 */
+	public BigDecimal getGameBetAmount() {
+		return gameBetAmount;
+	}
+
+	/**
+	 * Sets the amount to bet.
+	 *
+	 * @param gameBetAmount the amount to bet
+	 */
+	public void setGameBetAmount(BigDecimal gameBetAmount) {
+		this.gameBetAmount = gameBetAmount;
+	}
+
+	/**
+	 * Gets the amount to pay.
+	 *
+	 * @return amount to pay
+	 */
+	public BigDecimal getGamePaymentAmount() {
+		return gamePaymentAmount;
+	}
+
+	/**
+	 * Sets the amount to pay.
+	 *
+	 * @param gamePaymentAmount amount to pay
+	 */
+	public void setGamePaymentAmount(BigDecimal gamePaymentAmount) {
+		this.gamePaymentAmount = gamePaymentAmount;
+	}
+
+	/**
+	 * Gets the number of the credit card used to pay the bets.
+	 *
+	 * @return number of the credit card
+	 */
+	public String getGamePaymentCreditCardNumber() {
+		return gamePaymentCreditCardNumber;
+	}
+
+	/**
+	 * Sets the number of the credit card used to pay the bets.
+	 *
+	 * @param gamePaymentCreditCardNumber number of the credit card
+	 */
+	public void setGamePaymentCreditCardNumber(String gamePaymentCreditCardNumber) {
+		this.gamePaymentCreditCardNumber = gamePaymentCreditCardNumber;
+	}
+
+	/**
+	 * Gets the code of the credit card used to pay the bets.
+	 *
+	 * @return code of the credit card
+	 */
+	public String getGamePaymentCreditCardCode() {
+		return gamePaymentCreditCardCode;
+	}
+
+	/**
+	 * Sets the code of the credit card used to pay the bets.
+	 *
+	 * @param gamePaymentCreditCardCode code of the credit card
+	 */
+	public void setGamePaymentCreditCardCode(String gamePaymentCreditCardCode) {
+		this.gamePaymentCreditCardCode = gamePaymentCreditCardCode;
+	}
+
+	/**
+	 * Gets the expiration month the credit card used to pay the bets.
+	 *
+	 * @return expiration month the credit card
+	 */
+	public Integer getGamePaymentCreditCardExpirationMonth() {
+		return gamePaymentCreditCardExpirationMonth;
+	}
+
+	/**
+	 * Sets the expiration month the credit card used to pay the bets.
+	 *
+	 * @param gamePaymentCreditCardExpirationMonth expiration month the credit card
+	 */
+	public void setGamePaymentCreditCardExpirationMonth(Integer gamePaymentCreditCardExpirationMonth) {
+		this.gamePaymentCreditCardExpirationMonth = gamePaymentCreditCardExpirationMonth;
+	}
+
+	/**
+	 * Gets the expiration year the credit card used to pay the bets.
+	 *
+	 * @return expiration year the credit card
+	 */
+	public Integer getGamePaymentCreditCardExpirationYear() {
+		return gamePaymentCreditCardExpirationYear;
+	}
+
+	/**
+	 * Sets the expiration year the credit card used to pay the bets.
+	 *
+	 * @param gamePaymentCreditCardExpirationYear expiration year the credit card
+	 */
+	public void setGamePaymentCreditCardExpirationYear(Integer gamePaymentCreditCardExpirationYear) {
+		this.gamePaymentCreditCardExpirationYear = gamePaymentCreditCardExpirationYear;
+	}
+
+	/**
+	 * Gets the minimum expiration year for a valid credit card.
+	 *
+	 * @return minimum expiration year for a valid credit card
+	 */
+	public Integer getGamePaymentCreditCardExpirationYearMinimum() {
+		return gamePaymentCreditCardExpirationYearMinimum;
+	}
+
+	/**
+	 * Gets the maximum expiration year for a valid credit card.
+	 *
+	 * @return maximum expiration year for a valid credit card
+	 */
+	public Integer getGamePaymentCreditCardExpirationYearMaximum() {
+		return gamePaymentCreditCardExpirationYearMaximum;
 	}
 
 	/**
@@ -300,35 +459,96 @@ public class GameBean implements Serializable {
 		return getGameDAO().find(gameId);
 	}
 
+	/**
+	 * Saves (inserts or updates) the game.
+	 *
+	 * @return outcome: identifier of the view to navigate to
+	 */
 	public String saveGame() {
 
-		Team team1 = getTeamDAO().findById(gameTeam1);
-		Team team2 = getTeamDAO().findById(gameTeam2);
-		ZonedDateTime startTimeZoned = ZonedDateTime.of(gameStartTime, gameTimeZone);
+		try {
+			Team team1 = getTeamDAO().findById(gameTeam1);
+			Team team2 = getTeamDAO().findById(gameTeam2);
+			ZonedDateTime startTimeZoned = ZonedDateTime.of(gameStartTime, gameTimeZone);
 
-		Game saveGame = gameId != null ? getGameDAO().find(gameId) : null;
-		if (saveGame != null) {
-			saveGame.setTeam1(team1);
-			saveGame.setTeam2(team2);
-			saveGame.setStartTimeZoned(startTimeZoned);
-			getGameDAO().merge(saveGame);
+			Game game = gameId != null ? getGameDAO().find(gameId) : null;
+			if (game != null) {
+				game.setTeam1(team1);
+				game.setTeam2(team2);
+				game.setStartTimeZoned(startTimeZoned);
+				getGameDAO().merge(game);
 
-		} else {
-			saveGame = new Game(team1, team2, startTimeZoned);
-			getGameDAO().persist(saveGame);
+			} else {
+				game = new Game(team1, team2, startTimeZoned);
+				getGameDAO().persist(game);
+			}
+
+			return String.format("/players/game.xhtml?id=%d&faces-redirect=true", game.getId());
+
+		} catch (PersistenceException ex) {
+			MessageFactory.addError("ch.bfh.awebt.bookmaker.PERSISTENCE_ERROR");
 		}
 
-		return String.format("/players/game.xhtml?id=%d&faces-redirect=true", saveGame.getId());
+		return null;
 	}
 
+	/**
+	 * Deletes the game.
+	 *
+	 * @return outcome: identifier of the view to navigate to
+	 */
 	public String deleteGame() {
 
-		Game saveGame = gameId != null ? getGameDAO().find(gameId) : null;
-		if (saveGame != null) {
-			getGameDAO().remove(saveGame);
-			return "/home.xhtml"; //temporary
-		} else
-			return "Error no such Game"; //temporary     
+		try {
+			Game game = gameId != null ? getGameDAO().find(gameId) : null;
+			if (game != null) {
+				getGameDAO().remove(game);
+				return "/managers/upcoming-games.xhtml";
+
+			} else
+				MessageFactory.addWarning("ch.bfh.awebt.bookmaker.ITEM_NOT_FOUND_ERROR");
+
+		} catch (PersistenceException ex) {
+			MessageFactory.addError("ch.bfh.awebt.bookmaker.PERSISTENCE_ERROR");
+		}
+
+		return null;
+	}
+
+	/**
+	 * Shows the confirmation page for changing user bets.
+	 */
+	public void confirmUserBets() {
+
+		statusConfirmation = true;
+
+		UserBetDAO userBetDAO = getUserBetDAO();
+		gameBetAmount = BigDecimal.ZERO;
+		for (BetDTO bet: gameBets) {
+			UserBet userBet = userBetDAO.findByUserAndBet(loginBean.getUserId(), bet.getId());
+			gameBetAmount = gameBetAmount.add(bet.getStake().subtract(userBet != null ? userBet.getStake() : BigDecimal.ZERO));
+		}
+
+		User user = loginBean.getUser();
+		gamePaymentAmount = gameBetAmount.compareTo(user.getBalance()) > 0 ? gameBetAmount.subtract(user.getBalance()) : BigDecimal.ZERO;
+
+		gamePaymentCreditCardNumber = gamePaymentCreditCardCode = null;
+
+		LocalDate date = LocalDate.now();
+		gamePaymentCreditCardExpirationYearMinimum = gamePaymentCreditCardExpirationYear = date.getYear();
+		gamePaymentCreditCardExpirationYearMaximum = date.getYear() + 25;
+		gamePaymentCreditCardExpirationMonth = date.getMonthValue();
+	}
+
+	/**
+	 * Discards all changed user bets.
+	 */
+	public void cancelUserBets() {
+
+		statusConfirmation = false;
+
+		Integer gameId = this.gameId;
+		setGameId(gameId, true); //refresh all properties
 	}
 
 	/**
@@ -336,24 +556,49 @@ public class GameBean implements Serializable {
 	 */
 	public void saveUserBets() {
 
-		User user = loginBean.getUser();
-		BetDAO betDAO = getBetDAO();
-		UserBetDAO userBetDAO = getUserBetDAO();
+		try {
+			User user = loginBean.getUser();
+			BetDAO betDAO = getBetDAO();
+			UserBetDAO userBetDAO = getUserBetDAO();
 
-		for (BetDTO betDTO: gameBets) {
-			Bet bet = betDAO.find(betDTO.getId());
-			UserBet userBet = userBetDAO.findByUserAndBet(user, bet);
-			boolean persist = betDTO.getStake().compareTo(BigDecimal.ZERO) > 0;
+			gamePaymentAmount = gameBetAmount.compareTo(user.getBalance()) > 0 ? gameBetAmount.subtract(user.getBalance()) : BigDecimal.ZERO;
+			if (gamePaymentAmount.compareTo(BigDecimal.ZERO) > 0 && !confirmPayment()) {
+				MessageFactory.addWarning("ch.bfh.awebt.bookmaker.PAYMENT_ERROR");
+				return;
+			}
 
-			if (userBet != null)
-				if (persist) {
-					userBet.setStake(betDTO.getStake());
-					userBetDAO.merge(userBet);
-				} else
-					userBetDAO.remove(userBet);
+			user.deposit(gamePaymentAmount);
+			user.withdraw(gameBetAmount);
+			getUserDAO().merge(user);
 
-			else if (persist)
-				userBetDAO.persist(new UserBet(user, bet, betDTO.getStake()));
+			for (BetDTO betDTO: gameBets) {
+				Bet bet = betDAO.find(betDTO.getId());
+				UserBet userBet = userBetDAO.findByUserAndBet(user, bet);
+				boolean persist = betDTO.getStake().compareTo(BigDecimal.ZERO) > 0;
+
+				if (userBet != null)
+					if (persist) {
+						userBet.setStake(betDTO.getStake());
+						userBetDAO.merge(userBet);
+					} else
+						userBetDAO.remove(userBet);
+
+				else if (persist)
+					userBetDAO.persist(new UserBet(user, bet, betDTO.getStake()));
+			}
+
+			statusConfirmation = false;
+
+		} catch (PersistenceException ex) {
+			MessageFactory.addError("ch.bfh.awebt.bookmaker.PERSISTENCE_ERROR");
 		}
+	}
+
+	private boolean confirmPayment() {
+
+		LocalDate now = LocalDate.now();
+		LocalDate exp = LocalDate.of(gamePaymentCreditCardExpirationYear, gamePaymentCreditCardExpirationMonth, 1).plusMonths(1).minusDays(1);
+
+		return gamePaymentCreditCardCode != null && gamePaymentCreditCardCode != null && now.compareTo(exp) < 0; //implement actual payment process
 	}
 }
